@@ -1,11 +1,114 @@
 [build-system]
-requires = ["setuptools", "wheel"]
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
+
+[project]
+name = "pySPFM"
+description = "A python package for Paradigm Free Mapping (3dPFM and 3dMEPFM)."
+readme = "README.md"
+authors = [{name = "Eneko Uruñuela"}]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Information Analysis",
+    "License :: OSI Approved :: GNU Library or Lesser General Public License (LGPL)",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+]
+license = {file = "LICENSE"}
+requires-python = ">=3.8"
+dependencies = [
+    "dask",
+    "dask_jobqueue",
+    "distributed",
+    "jax",
+    "jaxlib",
+    "nibabel",
+    "nilearn==0.10.4",
+    "numpy>=1.22,<=1.26.4",
+    "pylops>=1.18.2",
+    "pyproximal>=0.4.0",
+    "pywavelets",
+    "scipy>=1.9.0",
+]
+dynamic = ["version"]
+
+[project.urls]
+Homepage = "https://github.com/Paradigm-Free-Mapping/pySPFM"
+Documentation = "https://www.pySPFM.readthedocs.io"
+
+[project.optional-dependencies]
+dev = ["pre-commit"]
+doc = [
+    "matplotlib",
+    "sphinx>=6.2.1",
+    "sphinx_copybutton",
+    "sphinx_rtd_theme>=1.2.2",
+    "sphinx-argparse",
+    "sphinxcontrib-bibtex",
+]
+
+tests = [
+    "codecov",
+    "coverage",
+    "flake8>=3.7",
+    "flake8-absolute-import",
+    "flake8-black",
+    "flake8-docstrings",
+    "flake8-isort",
+    "flake8-pyproject",
+    "flake8-unused-arguments",
+    "flake8-use-fstring",
+    "pep8-naming",
+    "pytest",
+    "pytest-cov>=4.0.0",
+    "pytest-console-scripts",
+    "requests",
+]
+
+# Aliases
+all = ["pySPFM[dev,doc,tests]"]
+
+[project.scripts]
+pySPFM = "pySPFM.workflows.pySPFM:_main"
+auc_to_estimates = "pySPFM.workflows.auc_to_estimates:_main"
+
+#
+# Hatch configurations
+#
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[tool.hatch.build.targets.sdist]
+exclude = [".git_archival.txt"]  # No longer needed in sdist
+
+[tool.hatch.build.targets.wheel]
+exclude = [
+    "pySPFM/tests/data",  # Large test data directory
+]
+
+##  The following two sections configure setuptools_scm in the hatch way
+
+[tool.hatch.version]
+source = "vcs"
+
+[tool.hatch.build.hooks.vcs]
+version-file = "pySPFM/_version.py"
+
+#
+# Developer tool configurations
+#
 
 [tool.black]
 line-length = 99
-target-version = ['py37']
+target-version = ['py38']
 include = '\.pyi?$'
 exclude = '''
+
 (
   /(
       \.eggs         # exclude a few common directories in the
@@ -13,17 +116,54 @@ exclude = '''
     | \.github
     | \.hg
     | \.pytest_cache
+    | \.testing_data_cache
     | _build
     | build
     | dist
   )/
-  | get_version.py
-  | versioneer.py
   | pySPFM/_version.py
-  | pySPFM/due.py
 )
 '''
 
 [tool.isort]
 profile = "black"
 multi_line_output = 3
+
+[tool.flake8]
+max-line-length = 99
+exclude = [
+    "*build/",
+    "pySPFM/_version.py",
+    "pySPFM/due.py",
+    "pySPFM/tests/*",
+]
+ignore = ["E203", "E402", "W503", "N802", "N806", "D403"]
+per-file-ignores = [
+    "*/__init__.py: F401",
+    "pySPFM/tests/*: D",
+]
+docstring-convention = "numpy"
+
+[tool.pydocstringformatter]
+write = true
+strip-whitespaces = true
+max-summary-lines = 1
+
+[tool.coverage.run]
+branch = true
+omit = [
+    "pySPFM/tests/*",
+    "*/__init__.py",
+    "pySPFM/_version.py",
+    "pySPFM/due.py",
+]
+
+[tool.coverage.report]
+# Regexes for lines to exclude from consideration
+exclude_lines = [
+    'raise NotImplementedError',
+    'warnings\.warn',
+]
+
+[tool.pytest]
+log_cli = true
