@@ -1,14 +1,14 @@
-"""
-Functions for parsers.
-"""
+"""Functions for parsers."""
+
 import argparse
 import os.path as op
 
 
 def check_hrf_value(string, is_parser=True):
     """
-    Check if argument is 'spm' or 'glover',
-    or is an existing '.1D' or '.txt' file.
+    Check if argument is 'spm' or 'glover'.
+
+    If it isn't, check if it is an existing '.1D' or '.txt' file.
     """
     if string == "spm" or string == "glover":
         return string
@@ -24,8 +24,20 @@ def check_hrf_value(string, is_parser=True):
 def is_valid_file(parser, arg):
     """
     Check if argument is existing file.
+
+    Parameters
+    ----------
+    parser : argparse.ArgumentParser
+        Argument parser.
+    arg : str
+        Argument to check.
+
+    Returns
+    -------
+    arg : str
+        Argument if it is an existing file.
     """
     if not op.isfile(arg) and arg is not None:
-        parser.error("The file {0} does not exist!".format(arg))
+        parser.error(f"The file {arg} does not exist!")
 
     return arg
