@@ -1,12 +1,16 @@
-# -*- coding: utf-8 -*-
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""
-pySPFM
-"""
+"""pySPFM."""
+import warnings
 
-from ._version import get_versions
+from pySPFM.__about__ import __copyright__, __credits__, __packagename__, __version__
 
-__version__ = get_versions()["version"]
+# cmp is not used, so ignore nipype-generated warnings
+warnings.filterwarnings("ignore", r"cmp not installed")
 
-del get_versions
+__all__ = [
+    "__copyright__",
+    "__credits__",
+    "__packagename__",
+    "__version__",
+]
