@@ -1,4 +1,5 @@
-"""HRF Matrix file."""
+"""Generate HRF matrix."""
+
 import logging
 import subprocess
 
@@ -56,7 +57,6 @@ class HRFMatrix:
         self.hrf_ : array_like
             A hemodynamic response function (HRF).
         """
-
         # Read custom HRF from file if self.model ends in .1D or .txt
         if self.model.endswith(".1D") or self.model.endswith(".txt"):
             hrf = np.loadtxt(self.custom)
@@ -68,38 +68,34 @@ class HRFMatrix:
                     "HRF is longer than the number of scans. "
                     "Please make sure that your custom HRF is not longer than the number of scans."
                 )
+        elif self.model == "spm":
+            hrf = spm_hrf(tr, oversampling=1, time_length=n_scans * tr)
+        elif self.model == "glover":
+            hrf = glover_hrf(tr, oversampling=1, time_length=n_scans * tr)
         else:
-            # Get HRF from nilearn
-            if self.model == "spm":
-                hrf = spm_hrf(tr, oversampling=1, time_length=n_scans * tr)
-            elif self.model == "glover":
-                hrf = glover_hrf(tr, oversampling=1, time_length=n_scans * tr)
-            else:
-                raise ValueError(
-                    "Model must be either 'spm', 'glover' or a custom '.1D' or '.txt' file, not %s"
-                    % self.model
-                )
-
-        # Calculate maximum HRF value
-        max_val = max(abs(hrf))
+            raise ValueError(
+                f"Model must be either 'spm', 'glover' or a custom '.1D' or '.txt' file, "
+                f"not {self.model}"
+            )
 
         # Generate HRF matrix
         hrf_mtx = hrf
         for i in range(n_scans - 1):
-            foo = np.append(np.zeros(i + 1), hrf[0 : (len(hrf) - i - 1)])
+            foo = np.append(np.zeros(i + 1), hrf[: len(hrf) - i - 1])
             hrf_mtx = np.column_stack((hrf_mtx, foo))
-
-        # Normalize HRF matrix
-        hrf_mtx = hrf_mtx / max_val
 
         # Concatenate and scale HRFs for multi-echo,
         # leave it as it is for single-echo.
         if len(self.te) > 1:
             # Add integrator if necessary
-            if self.block:
-                hrf_mtx_te = -self.te[0] * np.dot(hrf_mtx, np.tril(np.ones(n_scans)))
-            else:
-                hrf_mtx_te = -self.te[0] * hrf_mtx
+            hrf_mtx_te = (
+                -self.te[0] * np.dot(hrf_mtx, np.tril(np.ones(n_scans)))
+                if self.block
+                else -self.te[0] * hrf_mtx
+            )
+
+            # Calculate maximum HRF value
+            max_val = np.max(abs(hrf_mtx_te))
 
             # Concatenate and scale HRFs for multi-echo
             for teidx in range(len(self.te) - 1):
@@ -114,16 +110,19 @@ class HRFMatrix:
                 else:
                     hrf_mtx_te = np.vstack((hrf_mtx_te, -self.te[teidx + 1] * hrf_mtx))
 
-            self.hrf_ = hrf_mtx_te
+            # Normalize HRF
+            self.hrf_ = hrf_mtx_te / max_val
+        elif self.block:
+            hrf_non_norm = np.dot(hrf_mtx, np.tril(np.ones(n_scans)))
+
+            # Normalize HRF
+            self.hrf_ = hrf_non_norm / np.max(abs(hrf_non_norm))
         else:
-            if self.block:
-                self.hrf_ = np.dot(hrf_mtx, np.tril(np.ones(n_scans)))
-            else:
-                self.hrf_ = hrf_mtx
+            self.hrf_ = hrf_mtx / np.max(abs(hrf_mtx))
 
         return self
 
-    def hrf_afni(tr, hrf_model="SPMG1"):
+    def hrf_afni(self, hrf_model="SPMG1"):
         """Generate HRF with AFNI's 3dDeconvolve.
 
         Parameters
@@ -150,7 +149,8 @@ class HRFMatrix:
             hrf_command = (
                 "3dDeconvolve -x1D_stop -nodata %d %f -polort -1 -num_stimts 1 -stim_times 1 "
                 "'1D:0' '%s' -quiet -x1D stdout: | 1deval -a stdin: -expr 'a'"
-            ) % (dur_hrf, tr, hrf_model)
+                % (dur_hrf, self, hrf_model)
+            )
             hrf_tr_str = subprocess.check_output(
                 hrf_command, shell=True, universal_newlines=True
             ).splitlines()
@@ -164,7 +164,7 @@ class HRFMatrix:
 
         # Removes tail of zero samples
         while last_hrf_sample == 0:
-            hrf = hrf[0 : len(hrf) - 1]
+            hrf = hrf[:-1]
             last_hrf_sample = hrf[len(hrf) - 1]
 
         return hrf
